까먹고 수정 안 한 거 수정 ><

+++ src/main/kotlin/com/record/zooc/domain/entity/Textcomment.kt
@@ -3,6 +3,6 @@ import jakarta.persistence.DiscriminatorValue
 @DiscriminatorValue("Text")
 class Textcomment(
     content: String,
-    writerId: Int,
+    writerId: Long,
     recordId: Int,
 ) : Comment(content, writerId, recordId)
